requirements.txt
fastapi
uvicorn
python-multipart
chromadb
sentence-transformers
python-docx
pdfminer.six
python-dotenv
requests
